refactor: changed components file name
