+++ src/app/page.tsx
@@ -1,5 +1,1 @@
-import { redirect } from "next/navigation";
-
-export default async function Home() {
-  redirect("/auth/login");
-}
+export default async function Home() {}
