+++ src/app/professor/dashboard/disciplines-panel.tsx
